requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-redis==4.10.0
django-timezone-field==3.0
djangorestframework==3.9.4
idna==2.8
kombu==4.6.3
oauthlib==3.0.1
psycopg2-binary==2.8.3
PySocks==1.7.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tweepy==3.7.0
urllib3==1.25.3
vine==1.3.0
